Moved IDPSolverConfig to separate file
